+++ src/model/effects/EffectFactory.cpp
@@ -185,6 +185,8 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 			dynamic_cast<NetworkEffect *>(pEffect1);
 		BehaviorEffect * pBehaviorEffect1 =
 			dynamic_cast<BehaviorEffect *>(pEffect1);
+        ContinuousEffect * pContinuousEffect1 =
+            dynamic_cast<ContinuousEffect *>(pEffect1);
 
 		if (pNetworkEffect1)
 		{
@@ -200,7 +202,7 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 				pNetworkEffect2,
 				pNetworkEffect3);
 		}
-		else
+		else if (pBehaviorEffect1)
 		{
 			BehaviorEffect * pBehaviorEffect2 =
 				dynamic_cast<BehaviorEffect *>(pEffect2);
@@ -215,6 +217,22 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 				pBehaviorEffect2,
 				pBehaviorEffect3);
 		}
+        else
+        {
+            ContinuousEffect * pContinuousEffect2 =
+                dynamic_cast<ContinuousEffect *>(pEffect2);
+            ContinuousEffect * pContinuousEffect3 = 0;
+            if (pEffect3)
+            {
+                pContinuousEffect3 = dynamic_cast<ContinuousEffect *>(pEffect3);
+            }
+
+            pEffect = new ContinuousInteractionEffect(pEffectInfo,
+                pContinuousEffect1,
+                pContinuousEffect2,
+                pContinuousEffect3);
+        }
+        
 	}
 	else if (effectName == "density")
 	{
@@ -2019,6 +2037,13 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 			pEffect = new OutdegreeContinuousEffect(pEffectInfo, true);
 		}
 	}
+    else if (effectName == "outInBalance")
+    {
+        if (pContinuousData)
+        {
+            pEffect = new OutIndegreeBalanceContinuousEffect(pEffectInfo);
+        }
+    }
 	else if (effectName == "isolate")
 	{
 		pEffect = new IsolateEffect(pEffectInfo, true);
@@ -2191,6 +2216,10 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	{
 		pEffect = new AverageSimilarityInDist2Effect(pEffectInfo, false);
 	}
+    else if (effectName == "avSimVarAlt")
+    {
+        pEffect = new VarianceAlterSimilarityEffect(pEffectInfo, true, false, false);
+    }
 	else if (effectName == "avAlt")
 	{
 		if (pContinuousData)
@@ -2304,6 +2333,10 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	{
 		pEffect = new AverageAlterInDist2Effect(pEffectInfo, false, true);
 	}
+    else if (effectName == "varAlt")
+    {
+        pEffect = new VarianceAlterEffect(pEffectInfo);
+    }
 	else if (effectName == "behDenseTriads")
 	{
 		pEffect = new DenseTriadsBehaviorEffect(pEffectInfo);